+++ packages/eas-cli/src/credentials/manager/SelectPlatform.ts
@@ -3,6 +3,7 @@ import { DynamicConfigContextFn } from '../../commandUtils/context/DynamicProjec
 import { ExpoGraphqlClient } from '../../commandUtils/context/contextUtils/createGraphqlClient';
 import { selectPlatformAsync } from '../../platform';
 import { Actor } from '../../user/User';
+import { Client } from '../../vcs/vcs';
 import { CredentialsContextProjectInfo } from '../context';
 import { ManageAndroid } from './ManageAndroid';
 import { ManageIos } from './ManageIos';
@@ -11,6 +12,7 @@ export class SelectPlatform {
   constructor(
     public readonly actor: Actor,
     public readonly graphqlClient: ExpoGraphqlClient,
+    public readonly vcsClient: Client,
     public readonly analytics: Analytics,
     public readonly projectInfo: CredentialsContextProjectInfo | null,
     public readonly getDynamicPrivateProjectConfigAsync: DynamicConfigContextFn,